third_party/WebKit/LayoutTests/fast/speech/speech-input-result-list-not-enough-arguments-expected.txt

Tests for SpeechInputResultList' behavior with too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.event.results.item() threw exception TypeError: Failed to execute 'item' on 'SpeechInputResultList': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
